# swagger.yaml
basePath: /
consumes:
- application/json
definitions:
  Product:
    properties:
      description:
        type: string
        x-go-name: Description
      id:
        description: The id for this user
        format: int64
        minimum: 1
        type: integer
        x-go-name: ID
      name:
        type: string
        x-go-name: Name
      price:
        format: float
        type: number
        x-go-name: Price
      sku:
        type: string
        x-go-name: SKU
    required:
    - id
    type: object
    x-go-package: github.com/masha/WebServer/data
info:
  description: Documentation for Product API
  title: Product API
  version: 1.0.0
paths:
  /products:
    get:
      description: Returns a list of products
      operationId: listProducts
      responses:
        "200":
          $ref: '#/responses/productsResponse'
      tags:
      - products
    post:
      description: Return None
      operationId: postProduct
      parameters:
      - description: The product going to be post or put
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      tags:
      - products
    put:
      description: Return None
      operationId: putProduct
      parameters:
      - description: The product going to be post or put
        in: body
        name: Body
        required: true
        schema:
          $ref: '#/definitions/Product'
      tags:
      - products
  /products/{id}:
    delete:
      description: Returns a list of products
      operationId: deleteProduct
      parameters:
      - description: The id of the product to delete from the database
        format: int64
        in: path
        name: id
        required: true
        type: integer
        x-go-name: ID
      responses:
        "201":
          $ref: '#/responses/noContent'
      tags:
      - products
produces:
- application/json
responses:
  noContent:
    description: ""
  productsResponse:
    description: A list of products returns in response
    schema:
      items:
        $ref: '#/definitions/Product'
      type: array
schemes:
- http
swagger: "2.0"